update development environment
Update to newer libraries
and include linting
and testing.

The old code
was moved to `src-old` for now.
Once the code
is updated,
the `src-old` directory
can be removed.

The new code
doesn't do much
at the moment,
but the tests
and linters
all pass
with the new code.
The old code
will be ported over
to work with the new environment.