upgrade to React 16
Adds a requestAnimationFrame polyfill
for the unit tests,
since jsdom doesn't provide requestAnimationFrame,
and React 16 needs requestAnimationFrame.

Also fixes a new warning in React 16
about the `focusable` attribute
on the SVG icons
about using a boolean value
for a non-boolean attribute.
Using the string 'false'
instead of the boolean `false`
fixes this warning.