Display message when JavaScript is disabled
Most people
do not have JavaScript disabled,
and anyone who does
likely knows
they have disabled it
and how to re-enable it,
so let them know
that they need to
re-enabled JavaScript
to play this game.

Also, we have to skip
the Playwright test
(for now)
since the text within
the `noscript` element
can not be found
by Playwright
(even though it appears
in the Playwright UI
when running the test
with the UI).
This appears to be
a known issue
(at least,
there's a GitHub issue
about this problem).