Add game state and a way to initialize it
The game state currently controls
what screen the user sees.
The initialization methods allow us
to go directly to a certain screen
without navigating through the UI,
which is especially useful
for the end-to-end tests.

This adds vitest
so we can unit test
the game state methods.
It also adds valibot
so we can parse and validate
the given initial game state
which will disallow
invalid game states.