Use separate schemas for morning and day states
And combine them as a discriminated union
for the overall game state.

This will allow us
to add properties to the game state
that are valid for one screen
but not valid for another screen.

This change also requires
changing the value provided to `assertUnreachable`
from the game state's `screen` property
to the game state itself,
as now the entire game state
should have the `never` type
if none of the `switch` cases matched.