Use functions for text of news items
The `t` tag function from lingui needs to run
after lingui has been set up with the correct language,
so it has to be called from within a function
that runs after that setup happens.

Previously, we used a `generateNewsItems` function
to populate the news items in the `newsItemsByID` map
if a call to `getNewsItems` determined that the map was empty.
This does not work if the new `getNewsItem` function
is called before `getNewsItems` has been called.

Instead, we will use two different types of news items:

 - One news item type will use functions for the item text.
   The function, when called, returns the translated text.
 - The second type of news item has the translated text.

The data load will be the boundary
between the two types of news items.
So the UI code that gets its data from the loader
will always have the translated text.