Require file extensions in import statements
I have not been consistent
with either including or excluding file extensions
in import statements.
So, I made the decision
to always include them.

To help with this decision,
I have changed the TypeScript configuration
to use `"node16"`
for both the `module` and `moduleResolution`
compiler options.
This means that TypeScript will error
if an import excludes the file extension
(except for imports
from packages in node_modules).

I have also changed
the `resolve.extensions` vite configuration
to be an empty array.
This means that vite
will not try to add any file extensions
to imports that omit a file extension.
This means that vite will error at build time
if an import excludes the file extension
(again, except for imports
from packages in node_modules).