Use different TypeScript config for e2e tests
In the end-to-end tests,
we can use code
that we may not want to use
in the application code.
One reason may be
that we do not want to use
certain JavaScript features
in the application code,
due to the level of browser support
for the feature,
but we can use the feature
in the end-to-end tests.