Delay router creation until end of initialization
By creating and exporting the router at the top level of `router.ts`,
the router is created as soon as `router.ts` is included,
which is before we have everything initialized in `entry.tsx`
(such as loading and activating the translations).

Instead, have `router.ts` export a `getRouter` function
that will create and return the router.
We will call that function after we are done initializing everything.