Ensure error message is not visible after loading
Since we were only checking that the "Loading..." text was not visible
after the JavaScript loaded on the page,
it was possible for the test to pass
even though an error happened during loading,
because a loading error would hide the "Loading..." text
but display a loading error message.
So make sure we check that the "Loading..." text is not visible
AND that the loading error message is not visible.