Do not abuse invariant message callback
The second parameter to `invariant`
should be a message or a function that returns a message.
While we were returning a message,
we were also performing a side effect in our function
and if we are using a function to return a message,
I would like that function to be side-effect-free.

Instead, use a `try...catch` block to catch the invariant exception,
perform the side effect for the exceptional state,
then re-throw the invariant exception.
The effect is the same,
but now we do not have a message function with a side effect.