Replace createHashRouter with createBrowserRouter
I originally chose `createHashRouter`
to keep everything at the same URL,
because I was not sure if I could deploy this (whenever that happens)
as a static site with multiple URLs served up from the same static file.
I believe this is possible with Cloudflare Pages,
so I switching to `createBrowserRouter`.

This also means that we can use the URL from `request.url`
when calling `initializeGameStateFromURL` in `root.data.ts`,
since that URL has the search parameters.

As a side note,
I found out why the search parameters were missing from `request.url`
when we were using the hash router.
It was because the hash router parses the URL search parameters
from the hash portion of the URL.
So for a URL like `https://www.example.com/#?param=value`,
`request.url` would contain the search parameters
when using the hash router.