Change how news items are stored and retrieved
Call the `getArticleText` and `getFeedText` functions
when news items are retrieved
so we do not have to pass around news items
with untranslated text.

Remove the `getNewsItems` function
that returns all news items.
In its place,
explicitly get the news items we want
using `getNewsItem`.

This change requires us
to load the translation messages
and activate a locale
before the unit tests run.
We also provide a module definition
for the translation messages file
for TypeScript.