Ensure articles reference items in newsItems array
The plan is to have all the news items for a given day
listed in the `newsItems` array.
Articles added to the paper
will reference the news items
in that `newsItems` array.
That way, when sending data from the loader,
when can send the full news item object
for each item in the `newsItems` array,
and reference those news items by ID
in the `newsItem` property of each article.