Use a single tsconfig file
While I liked the separate tsconfig files
for the application files and the test files,
since it allowed me to specify different options,
I ran into some problems with the unit test files
not picking up custom matcher types
since they were in separate TypeScript projects
(due to the separate tsconfig files).

I could move the unit test files
to be within the `tests` directory,
but I like co-locating the unit test file
with the file (and code) it is testing.

I also tried including the unit test files
in the test tsconfig project
instead of in the root tsconfig project,
but then I ran into an issue
with the typescript-eslint plugin
complaining that the project service
could not find the unit test files.