Add status text to day loader's invariantResponse
The error page code will display the status text.
`invariantResponse` does not default the status text to a value.
The default value of `statusText` on a `Response` object,
which is what `invariantResponse` throws,
is an empty string.

It is unlikely that a user would see the error page
due to `invariantResponse` throwing a `Response` object
in the day route's loader,
since the root loader would see the state mismatch
and redirect the user
before the day route's loader executed.