Update workaround for ReadonlyMap loader data
In React Router v7.3.0,
the `unstable_SerializesTo` type was changed,
making the `unstable__ReactRouter_SerializesTo` brand required,
whereas before it was optional.
This means that the `ReadonlyMap`
needs to be cast to `unknown`
before being cast to `unstable_SerializesTo<ReadonlyMap>`.

However, casting to `unknown`
means we can no longer check that the `ReadonlyMap`
contains keys and values that serialize,
since we lose the `ReadonlyMap` type information
after casting it to `unknown`.
To address this,
we have a `toSerializableReadonlyMap` function
that uses a `SerializableReadonlyMap` type
that ensures that the given parameter
is a `ReadonlyMap` with serializable keys and values.
With the argument checked at the function call site,
we can safely cast the argument to `unknown`
and then to `unstable_SerializesTo`.