Build as part of validate script
When Playwright was running the `npm run preview` command,
triggering the build via the `prepreview` script,
we were getting a minified file size warning for the `index.js` asset,
as the minified file was over 500 kB in size.
However, run building directly via `npm run build`,
the minified file was about 291 kB.
I'm guessing that some environment variable,
or combination of environment variables,
that Playwright sets when executing
was causing the built file to be almost two times larger.