Disallow direct imports of @epic-web/invariant
This insures that we will receive an ESLint error
if we forget to import from `#app/invariant.ts`
instead of `@epic-web/invariant`.

This rule does not apply to `#app/invariant.ts`,
since that file needs to import from `@epic-web/invariant`
so we can wrap the invariant functions
(and export `InvariantError`,
in case we ever want/need to use that).