Add different news item ID schema implementation
It was nice to have the news item ID type
limited to only the exact IDs we were using,
but it came with a large tradeoffs that I did not like:
The news item ID type depended on the news item data.
If we were to ever change how the news item data was defined or stored,
such as moving the news item data to a database,
it would be a lot more difficult to get all the possible IDs
so we could define the picklist.

This implementation does not use a picklist,
so it does not depend on the news item data for its definition.
That makes it easier to import the schema, type,
or the associated helpers
without also causing an import of all the news item data.

We are also using a "branded" type for the news item ID,
so only strings that have been validated
as having the correct characters and length
can be considered a possible news item ID.
Even after that validation,
it is possible that the news item does not exist,
but at least we know that it is possible
that an item with that ID could exist.