Render a single form for each news item
The form for each news item will always be rendered,
so we can have submit buttons in multiple places on the page
that reference the news item's form via the `form` attribute.
This will be helpful when we add more actions involving news items,
so each of those actions does not have to render its own form.