Use day for attribute and property name
Using `day-number` for the attribute name
and `dayNumber` for the property name
could lead to some potential issues
or confusion.

Using `day` for both makes it clearer
that the attribute and property
refer to the same thing.
We do lose a little context
that this should be the day number,
but I think the trade off is worth it.