Change stateincontrol attr/prop to government
The default state of the component
should be the starting, day 1 state.
That means, without any attributes or properties set,
it should display the day 1 state.

With the `stateincontrol` boolean attribute,
it would need to be `true` for the day 1 state.
That would mean the attribute would need to be present.

So I'm deviating from the original game's code
and using a `government` attribute
with an enumerated value.