Do not validate values saved to attributes
For built-in DOM elements like `<input>`,
property setters do not validate the value.
Instead, they pass that value directly to the attribute
(for those properties that mirror attribute values, at least).
When the property getter is used to get the property's value,
the attribute's value is then validated.
If the attribute contains an invalid value,
a default value is returned.
If the attribute contains a floating point number
when it should be an integer value,
the integer portion of the floating point number is returned.

The change aligns the property setter behavior of `<morning-state>`
with that of built-in DOM elements
(as far as I can tell, anyway).